minor, mostly cosmetic changes and test moves
